requirements.txt
cffi==1.15.1
click==8.1.3
colorama==0.4.6
-e git+https://github.com/d-exponent/cli_todo_app.git@74c7c63e6f36c700157b58163e70d320dbcd3121#egg=eclipse_todo
greenlet==2.0.2
iniconfig==2.0.0
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mdurl==0.1.2
mysql-connector-python==8.0.33
mysqlclient==2.2.0
numpy==1.24.3
packaging==23.1
pandas==2.0.2
pluggy==1.0.0
protobuf==3.20.3
psycopg2==2.9.6
pycparser==2.21
Pygments==2.15.1
pytest==7.3.2
python-dateutil==2.8.2
pytz==2023.3
rich==13.4.1
shellingham==1.5.0.post1
six==1.16.0
SQLAlchemy==2.0.15
typer==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
